Migrate names of builtins-replacements methods
Migrate names of builtins-replacements methods. Add also some
overloading interfaces of those builtins procedures.

Side effects:

Refactor the sources structure: to sanitize the overloading interfaces
of builtins procedures these have be placed outside the definition of
the string class, thus a new module is added, namely the string_t.f90
module and the old stringifor.f90 module is used to define the
overlading interfaces and exposes objects to the user.