Fix issue#16
Change intent of an host associated variable in `split_last_token` in
`split_chunked`: intent(out) is not standard conforming with pure
attribute.

The NAG Fortran Compiler notices that:

Error: src/StringiFor/stringifor_string_t.F90, line 1551: TOKENS cannot
be INTENT(OUT) actual arg in PURE procedure SPLIT_LAST_TOKEN

SPLIT_LAST_TOKEN is a PURE internal subroutine of PURE subroutine
SPLIT_CHUNKED which has a TOKENS INTENT(OUT) dummy argument. TOKENS is
accessed in SPLIT_LAST_TOKEN by host association and is used as the TO
actual argument of MOVE_ALLOC which is INTENT(OUT). That means TOKENS is
used in a "variable definition context" which is expressly not allowed
for host-association accessed entities.

The restriction is C1272 in "12.6 Pure procedures" of the 2003 Fortran
Standard. The Final Committee Draft (FCD) of the new Standard (Fortran
2003) is available as PDF from
http://www.dkuug.dk/jtc1/sc22/open/n3661.pdf

C1272
In a pure subprogram any designator with a base object that is [...]
accessed by [...] use association [...], shall not be used in the
following contexts:
(1) In a variable deﬁnition context(16.5.7);

16.5.7 Variable deﬁnition context
Some variables are prohibited from appearing in a syntactic context that
would imply deﬁnition or un- deﬁnition of the variable (5.1.2.7,
5.1.2.12, 12.6). The following are the contexts in which the appearance
of a variable implies such deﬁnition or undeﬁnition of the variable:
[...]
(10) An actual argument in a reference to a procedure with an explicit
interface if the associated dummy argument has the INTENT(OUT) or
INTENT(INOUT) attribute

Side effects:

Nothing