Disable len overloading
Disable len overloading

Why:

Intel Compiler (18.0.0) does not accept the "len" overloading because it
is used into variables initialization into declaration and it does not
understand that the overloading procedures are "pure"... it is probably
necessary to reduce the error in order to fill a bug report. No time to
do it now.

This change addresses the need by:

Disable the len overloading interface.

Side effects:

It is not possible to perform "len(string)".

Must be re-introduced in the near future.